change type of coordinates to double

+++ решения/cs/Methods/Lab05.cs
@@ -487,10 +487,10 @@ namespace Methods
 
         public class Point
         {
-            public readonly int x;
-            public readonly int y;
+            public readonly double x;
+            public readonly double y;
 
-            public Point(int x, int y)
+            public Point(double x, double y)
             {
                 this.x = x;
                 this.y = y;
